feat: 完成迭代优化并更新测试报告
- 修复错误分类器模块测试 (21/21 passed)
- 修复补丁应用模块测试 (20/20 passed)
- 修复路径占位符问题
- 添加完整测试报告文档
- 更新README添加测试状态
- 删除论文相关文档(敏感信息)

测试状态: 436 passed, 2 warnings

+++ examples/calculator.py
@@ -25,7 +25,7 @@ def subtract(a: float, b: float) -> float:
     Returns:
         a - b 的计算结果。
     """
-    return a - b
+    return float(a - b)
 
 
 def factorial(n: int) -> int:
@@ -61,7 +61,7 @@ def add(a: float, b: float) -> float:
     Returns:
         a + b 的计算结果。
     """
-    return a + b
+    return float(a + b)
 
 
 def divide(a: float, b: float) -> float:
@@ -81,7 +81,7 @@ def divide(a: float, b: float) -> float:
     # 除零保护：避免 ZeroDivisionError，主动抛出 ValueError 以便测试捕获
     if b == 0:
         raise ValueError("除数不能为零")
-    return a / b
+    return float(a / b)
 
 
 def multiply(a: float, b: float) -> float:
@@ -94,4 +94,4 @@ def multiply(a: float, b: float) -> float:
     Returns:
         a * b 的计算结果。
     """
-    return a * b
+    return float(a * b)
